requirements.txt
py-cord
python-dotenv
audioop-lts